I uploaded a release candidate for both the HTML and the CSS temaplate
I've integrated Janies additions to the template and created a minified version of the CSS with very minimal changes outside of the wrapper-jiraNum id this will need to be tested and approve by Jay. I made an aditional change in the template to make it easier for team to download and implement, the starter template is referencing the local copy of the CSS that way they can make the changes needed before uploading to the server.